+++ src/managers/run-strategies/reuse-last-initialized.js
@@ -46,7 +46,7 @@ module.exports = classFrom(Base, {
         });
     },
 
-    getRun: function (runService, userSession, runIdInSession, options) {
+    getRun: function (runService, userSession, runSession, options) {
         var filter = injectFiltersFromSession(this.options.flag, userSession);
         var me = this;
         return runService.filter(filter, {